requirements.txt
# AI示波器控制系统依赖库
# 项目版本: 1.0.0
# Python版本要求: >=3.8

# ================================
# 核心框架依赖
# ================================

# UI框架 - PySide6+QML
PySide6>=6.5.0,<7.0.0
PySide6-Essentials>=6.5.0
PySide6-Addons>=6.5.0

# AI框架 - LangChain
langchain>=0.1.0,<0.3.0
langchain-community>=0.1.0,<0.3.0
langchain-core>=0.1.0,<0.3.0
openai>=1.0.0,<2.0.0

# Actor系统 - 多线程消息交互
pykka>=4.0.0,<5.0.0

# ================================
# 数据处理和分析
# ================================

# 科学计算
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<2.0.0
pandas>=2.0.0,<3.0.0

# 可视化
matplotlib>=3.7.0,<4.0.0
plotly>=5.15.0,<6.0.0

# ================================
# 设备控制和通信
# ================================

# 串口通信
pyserial>=3.5,<4.0
pyusb>=1.2.1,<2.0.0

# 网络通信
requests>=2.31.0,<3.0.0
httpx>=0.25.0,<1.0.0

# ================================
# 配置和文件处理
# ================================

# 配置文件
pyyaml>=6.0,<7.0
toml>=0.10.2,<1.0.0
python-dotenv>=1.0.0,<2.0.0

# 文档和模板
jinja2>=3.1.0,<4.0.0
markdown>=3.4.0,<4.0.0

# ================================
# 开发和调试工具
# ================================

# 代码格式化和检查
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
pylint>=2.17.0,<3.0.0

# 测试框架
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-asyncio>=0.21.0,<1.0.0

# 类型检查
mypy>=1.5.0,<2.0.0

# ================================
# 系统和工具
# ================================

# Windows系统支持
pywin32>=306;sys_platform=="win32"
colorama>=0.4.6,<1.0.0

# 时间处理
python-dateutil>=2.8.2,<3.0.0

# 进度条和美化
tqdm>=4.65.0,<5.0.0
rich>=13.4.0,<14.0.0

# 加密和安全
cryptography>=41.0.0,<42.0.0
